refactor shared functions for typed list and dict
The typed list and dict implementations ended up sharing some utilities.
This commit moves and refactors the ones that were easy enough to
generlize into a shared module called typedobjectutils.py and re-wires
all the imports.
