+++ numba/dictobject.py
@@ -9,7 +9,6 @@ from llvmlite import ir
 
 from numba import cgutils
 from numba import _helperlib
-from numba.targets.registry import cpu_target
 
 from numba.extending import (
     overload,
@@ -29,10 +28,17 @@ from numba.types import (
     DictIteratorType,
     Type,
 )
-from numba.typeconv import Conversion
 from numba.targets.imputils import impl_ret_borrowed, RefType
 from numba.errors import TypingError
 from numba import typing
+from numba.typedobjectutils import (_as_bytes,
+                                    _cast,
+                                    _nonoptional,
+                                    _sentry_safe_cast_default,
+                                    _get_incref_decref,
+                                    _get_equal,
+                                    _container_get_data,
+                                    )
 
 
 ll_dict_type = cgutils.voidptr_t
@@ -201,81 +207,6 @@ def _imp_dtor(context, module):
     return fn
 
 
-def _dict_get_data(context, builder, dict_ty, d):
-    """Helper to get the C dict pointer in a numba dict.
-    """
-    ctor = cgutils.create_struct_proxy(dict_ty)
-    dstruct = ctor(context, builder, value=d)
-    return dstruct.data
-
-
-def _as_bytes(builder, ptr):
-    """Helper to do (void*)ptr
-    """
-    return builder.bitcast(ptr, cgutils.voidptr_t)
-
-
-def _sentry_safe_cast(fromty, toty):
-    """Check and raise TypingError if *fromty* cannot be safely cast to *toty*
-    """
-    tyctxt = cpu_target.typing_context
-    by = tyctxt.can_convert(fromty, toty)
-    if by is None or by > Conversion.safe:
-        if isinstance(fromty, types.Integer) and isinstance(toty, types.Integer):
-            # Accept if both types are ints
-            return
-        if isinstance(fromty, types.Integer) and isinstance(toty, types.Float):
-            # Accept if ints to floats
-            return
-        if isinstance(fromty, types.Float) and isinstance(toty, types.Float):
-            # Accept if floats to floats
-            return
-        raise TypingError('cannot safely cast {} to {}'.format(fromty, toty))
-
-
-def _sentry_safe_cast_default(default, valty):
-    """Similar to _sentry_safe_cast but handle default value.
-    """
-    # Handle default values
-    # TODO: simplify default values; too many possible way to spell None
-    if default is None:
-        return
-    if isinstance(default, (types.Omitted, types.NoneType)):
-        return
-    return _sentry_safe_cast(default, valty)
-
-
-@intrinsic
-def _cast(typingctx, val, typ):
-    """Cast *val* to *typ*
-    """
-    def codegen(context, builder, signature, args):
-        [val, typ] = args
-        context.nrt.incref(builder, signature.return_type, val)
-        return val
-    # Using implicit casting in argument types
-    casted = typ.instance_type
-    _sentry_safe_cast(val, casted)
-    sig = casted(casted, typ)
-    return sig, codegen
-
-
-@intrinsic
-def _nonoptional(typingctx, val):
-    """Typing trick to cast Optional[T] to T
-    """
-    if not isinstance(val, types.Optional):
-        raise TypeError('expected an optional')
-
-    def codegen(context, builder, sig, args):
-        context.nrt.incref(builder, sig.return_type, args[0])
-        return args[0]
-
-    casted = val.type
-    sig = casted(casted)
-    return sig, codegen
-
-
 @intrinsic
 def _dict_new_minsize(typingctx, keyty, valty):
     """Wrap numba_dict_new_minsize.
@@ -317,82 +248,6 @@ def _dict_new_minsize(typingctx, keyty, valty):
     return sig, codegen
 
 
-def _get_incref_decref(context, module, datamodel):
-    assert datamodel.contains_nrt_meminfo()
-
-    fe_type = datamodel.fe_type
-    data_ptr_ty = datamodel.get_data_type().as_pointer()
-    refct_fnty = ir.FunctionType(ir.VoidType(), [data_ptr_ty])
-    incref_fn = module.get_or_insert_function(
-        refct_fnty,
-        name='.numba_dict_incref${}'.format(fe_type),
-    )
-    builder = ir.IRBuilder(incref_fn.append_basic_block())
-    context.nrt.incref(builder, fe_type, builder.load(incref_fn.args[0]))
-    builder.ret_void()
-
-    decref_fn = module.get_or_insert_function(
-        refct_fnty,
-        name='.numba_dict_decref${}'.format(fe_type),
-    )
-    builder = ir.IRBuilder(decref_fn.append_basic_block())
-    context.nrt.decref(builder, fe_type, builder.load(decref_fn.args[0]))
-    builder.ret_void()
-
-    return incref_fn, decref_fn
-
-
-def _get_equal(context, module, datamodel):
-    assert datamodel.contains_nrt_meminfo()
-
-    fe_type = datamodel.fe_type
-    data_ptr_ty = datamodel.get_data_type().as_pointer()
-
-    wrapfnty = context.call_conv.get_function_type(types.int32, [fe_type, fe_type])
-    argtypes = [fe_type, fe_type]
-
-    def build_wrapper(fn):
-        builder = ir.IRBuilder(fn.append_basic_block())
-        args = context.call_conv.decode_arguments(builder, argtypes, fn)
-
-        sig = typing.signature(types.boolean, fe_type, fe_type)
-        op = operator.eq
-        fnop = context.typing_context.resolve_value_type(op)
-        fnop.get_call_type(context.typing_context, sig.args, {})
-        eqfn = context.get_function(fnop, sig)
-        res = eqfn(builder, args)
-        intres = context.cast(builder, res, types.boolean, types.int32)
-        context.call_conv.return_value(builder, intres)
-
-    wrapfn = module.get_or_insert_function(
-        wrapfnty,
-        name='.numba_dict_key_equal.wrap${}'.format(fe_type)
-    )
-    build_wrapper(wrapfn)
-
-    equal_fnty = ir.FunctionType(ir.IntType(32), [data_ptr_ty, data_ptr_ty])
-    equal_fn = module.get_or_insert_function(
-        equal_fnty,
-        name='.numba_dict_key_equal${}'.format(fe_type),
-    )
-    builder = ir.IRBuilder(equal_fn.append_basic_block())
-    lhs = datamodel.load_from_data_pointer(builder, equal_fn.args[0])
-    rhs = datamodel.load_from_data_pointer(builder, equal_fn.args[1])
-
-    status, retval = context.call_conv.call_function(
-        builder, wrapfn, types.boolean, argtypes, [lhs, rhs],
-    )
-    with builder.if_then(status.is_ok, likely=True):
-        with builder.if_then(status.is_none):
-            builder.ret(context.get_constant(types.int32, 0))
-        retval = context.cast(builder, retval, types.boolean, types.int32)
-        builder.ret(retval)
-    # Error out
-    builder.ret(context.get_constant(types.int32, -1))
-
-    return equal_fn
-
-
 @intrinsic
 def _dict_set_method_table(typingctx, dp, keyty, valty):
     """Wrap numba_dict_set_method_table
@@ -429,9 +284,9 @@ def _dict_set_method_table(typingctx, dp, keyty, valty):
 
         dm_key = context.data_model_manager[keyty.instance_type]
         if dm_key.contains_nrt_meminfo():
-            equal = _get_equal(context, builder.module, dm_key)
+            equal = _get_equal(context, builder.module, dm_key, 'dict')
             key_incref, key_decref = _get_incref_decref(
-                context, builder.module, dm_key,
+                context, builder.module, dm_key, 'dict'
             )
             builder.store(
                 builder.bitcast(equal, key_equal_ptr.type.pointee),
@@ -449,7 +304,7 @@ def _dict_set_method_table(typingctx, dp, keyty, valty):
         dm_val = context.data_model_manager[valty.instance_type]
         if dm_val.contains_nrt_meminfo():
             val_incref, val_decref = _get_incref_decref(
-                context, builder.module, dm_val,
+                context, builder.module, dm_val, 'dict'
             )
             builder.store(
                 builder.bitcast(val_incref, val_incref_ptr.type.pointee),
@@ -492,7 +347,7 @@ def _dict_insert(typingctx, d, key, hashval, val):
         # TODO: the ptr_oldval is not used.  needed for refct
         ptr_oldval = cgutils.alloca_once(builder, data_val.type)
 
-        dp = _dict_get_data(context, builder, td, d)
+        dp = _container_get_data(context, builder, td, d)
         status = builder.call(
             fn,
             [
@@ -525,7 +380,7 @@ def _dict_length(typingctx, d):
         fn = builder.module.get_or_insert_function(fnty, name='numba_dict_length')
         [d] = args
         [td] = sig.args
-        dp = _dict_get_data(context, builder, td, d)
+        dp = _container_get_data(context, builder, td, d)
         n = builder.call(fn, [dp])
         return n
 
@@ -547,7 +402,7 @@ def _dict_dump(typingctx, d):
         )
         [td] = sig.args
         [d] = args
-        dp = _dict_get_data(context, builder, td, d)
+        dp = _container_get_data(context, builder, td, d)
         fn = builder.module.get_or_insert_function(fnty, name='numba_dict_dump')
 
         builder.call(fn, [dp])
@@ -582,7 +437,7 @@ def _dict_lookup(typingctx, d, key, hashval):
         ll_val = context.get_data_type(td.value_type)
         ptr_val = cgutils.alloca_once(builder, ll_val)
 
-        dp = _dict_get_data(context, builder, td, d)
+        dp = _container_get_data(context, builder, td, d)
         ix = builder.call(
             fn,
             [
@@ -634,7 +489,7 @@ def _dict_popitem(typingctx, d):
         ptr_key = cgutils.alloca_once(builder, dm_key.get_data_type())
         ptr_val = cgutils.alloca_once(builder, dm_val.get_data_type())
 
-        dp = _dict_get_data(context, builder, td, d)
+        dp = _container_get_data(context, builder, td, d)
         status = builder.call(
             fn,
             [
@@ -677,7 +532,7 @@ def _dict_delitem(typingctx, d, hk, ix):
 
         fn = builder.module.get_or_insert_function(fnty, name='numba_dict_delitem')
 
-        dp = _dict_get_data(context, builder, td, d)
+        dp = _container_get_data(context, builder, td, d)
         status = builder.call(fn, [dp, hk, ix])
         return status
 
@@ -1092,7 +947,7 @@ def impl_iterable_getiter(context, builder, sig, args):
     pstate = cgutils.alloca_once(builder, state_type, zfill=True)
     it.state = _as_bytes(builder, pstate)
 
-    dp = _dict_get_data(context, builder, iterablety.parent, it.parent)
+    dp = _container_get_data(context, builder, iterablety.parent, it.parent)
     builder.call(fn, [it.state, dp])
     return impl_ret_borrowed(
         context,
@@ -1126,7 +981,7 @@ def impl_dict_getiter(context, builder, sig, args):
     it.state = _as_bytes(builder, pstate)
     it.parent = d
 
-    dp = _dict_get_data(context, builder, iterablety.parent, args[0])
+    dp = _container_get_data(context, builder, iterablety.parent, args[0])
     builder.call(fn, [it.state, dp])
     return impl_ret_borrowed(
         context,
