+++ numba/tests/test_dictobject.py
@@ -15,6 +15,7 @@ from numba import njit, utils
 from numba import int32, int64, float32, float64, types
 from numba import dictobject, typeof
 from numba.typed import Dict
+from numba.typedobjectutils import _sentry_safe_cast
 from numba.utils import IS_PY3
 from numba.errors import TypingError
 from .support import TestCase, MemoryLeakMixin, unittest
@@ -896,11 +897,11 @@ class TestDictObject(MemoryLeakMixin, TestCase):
 
 class TestDictTypeCasting(TestCase):
     def check_good(self, fromty, toty):
-        dictobject._sentry_safe_cast(fromty, toty)
+        _sentry_safe_cast(fromty, toty)
 
     def check_bad(self, fromty, toty):
         with self.assertRaises(TypingError) as raises:
-            dictobject._sentry_safe_cast(fromty, toty)
+            _sentry_safe_cast(fromty, toty)
         self.assertIn(
             'cannot safely cast {fromty} to {toty}'.format(**locals()),
             str(raises.exception),
